Remove behavior in daemon where we crash the process if any thread hasn't heartbeated within a certain time period
Summary:
Right now the dagster daemon has a feature where if a given daemon doesn't heartbeat (by yielding within its loop) within a certain period of time, we assume that it is hanging and crashed the process. The thinking was that we wanted to fail loudly if a process hanged and give the daemon a chance to recover.

In practice, hangs like this are rare (and can be diagnosed using py-spy). However, it is quite easy to inadvertently write slightly slower than normal code or forget to yield, hit this logic, and crash the process. The backfill daemon has run into this in the past, and the asset daemon can easily run into a situation with larger asset graphs where the initial attempt to generate materializations takes a few minutes to run.

So this PR removes that behavior. We will still detect and kill the process if a thread dies, we just don't try to determine whether the thread is hanging by looking for heartbeats.

Test Plan: BK, modified tests, deploy a daemon locally and verify that a segfault in one thread still kills the whole process.